Listener vs Pedidos Automaticos

+++ brm-service.api/src/main/java/br/com/brm/scp/api/dto/request/PedidoRequestDTO.java
@@ -26,7 +26,7 @@ public class PedidoRequestDTO implements Serializable {
 
 	private String destino;
 
-	private Integer quantidade;
+	private Long quantidade;
 
 	private PedidoStatus status;
 
@@ -77,11 +77,11 @@ public class PedidoRequestDTO implements Serializable {
 		this.destino = destino;
 	}
 
-	public Integer getQuantidade() {
+	public Long getQuantidade() {
 		return quantidade;
 	}
 
-	public void setQuantidade(Integer quantidade) {
+	public void setQuantidade(Long quantidade) {
 		this.quantidade = quantidade;
 	}
