+++ brm-service.api/src/main/java/br/com/brm/scp/api/service/SkuService.java
@@ -33,7 +33,7 @@ public interface SkuService {
 
 	Collection<SkuResponseDTO> chain(String idItem) throws SkuNotFoundException;
 
-	void addEstoque(String id, Integer quantidade) throws SkuNotFoundException, SkuNotSuchMuchQuantityException;
+	void addEstoque(String id, Long quantidade) throws SkuNotFoundException, SkuNotSuchMuchQuantityException;
 
 	void estoqueSeguranca(String id, Double es) throws SkuNotFoundException;
