+++ brm-service.impl/src/main/java/br/com/brm/scp/api/service/document/PedidoDocument.java
@@ -31,7 +31,7 @@ public class PedidoDocument implements Serializable {
 
 	private String destino; // ID SKU ORIGEM
 
-	private Integer quantidade;
+	private Long quantidade;
 
 	private PedidoStatus status;
 
@@ -86,11 +86,11 @@ public class PedidoDocument implements Serializable {
 		this.destino = destino;
 	}
 
-	public Integer getQuantidade() {
+	public Long getQuantidade() {
 		return quantidade;
 	}
 
-	public void setQuantidade(Integer quantidade) {
+	public void setQuantidade(Long quantidade) {
 		this.quantidade = quantidade;
 	}
